refactor(kilo-chat): split message input responsibilities
